+++ CHANGES.rst
@@ -42,10 +42,10 @@ New features
   :pr:`1203` by :user:`Reshama Shaikh <reshamas>` and
   :user:`Vincent Maladiere <Vincent-Maladiere>`.
 
-- The :class:`SimpleCleaner` is now available. This transformer is a lightweight
+- ⚠ The :class:`SimpleCleaner` was renamed to :class:`Cleaner` in 0.5.3.
+  The :class:`SimpleCleaner` is now available. This transformer is a lightweight
   pre-processor that applies some of the transformations applied by the :class:`TableVectorizer`, with a simpler
   interface. :pr:`1266` by :user:`Riccardo Cappuzzo<rcap107>` and :user:`Jerome Dockes <jeromedockes>` .
-  ⚠ The :class:`SimpleCleaner` was renamed to :class:`Cleaner` in 0.5.3.
 Changes
 -------