+++ CHANGES.rst
@@ -6,7 +6,7 @@ Release history
 
 .. currentmodule:: skrub
 
-Release 0.5.2
+Release 0.5.3
 ===================
 
 Changes